feat: stockfish integration
