+++ .gitignore
@@ -7,3 +7,4 @@ venv/
 .env
 .next/
 out/
+stockfish/stockfish-macos
